Fix Makefile for directory changes
The python scripts were moved into a subdirectory of the project,
lean_schema/ . The two fixes required in the makefile for this are:

1) The Makefile target lean_schema had to be declared as PHONY since it
was looking at the lean_schema/ directory for changes. This caused the
target to never run, as it was never updated.

2) The python script paths had to be changed to refer to the script
within the lean_schema/ directory.

Other targets were also added to PHONY to prevent future issues.